increase cooldowns and fix dbcontext in emon

+++ StreamSchedule/Commands/Helps.cs
@@ -8,7 +8,7 @@ internal class Helps : Command
     internal override string Call => "helps";
     internal override Privileges MinPrivilege => Privileges.None;
     internal override string Help => "show command help: [command name]";
-    internal override TimeSpan Cooldown => TimeSpan.FromSeconds((int)Cooldowns.Short);
+    internal override TimeSpan Cooldown => TimeSpan.FromSeconds((int)Cooldowns.Long);
     internal override Dictionary<string, DateTime> LastUsedOnChannel { get; set; } = [];
     internal override string[]? Arguments => null;
