+++ StreamSchedule/Commands/Roll.cs
@@ -7,7 +7,7 @@ internal class Roll : Command
     internal override string Call => "roll";
     internal override Privileges MinPrivilege => Privileges.None;
     internal override string Help => "roll a die [amount(max 100_000)]d[sides(max 100_000)], -v for details (up to 50 results)";
-    internal override TimeSpan Cooldown => TimeSpan.FromSeconds((int)Cooldowns.Short);
+    internal override TimeSpan Cooldown => TimeSpan.FromSeconds((int)Cooldowns.Long);
     internal override Dictionary<string, DateTime> LastUsedOnChannel { get; set; } = [];
     internal override string[] Arguments => ["v"];
